feat! : migrate to .net 5.0
* migrate project to .net core sdk
* migrate to service collection and dependency injection
* convert UserSettings to settings.json file type
* separate to a common project and windows specific one
* removed unused files
* refactor "TickerController"
* refactor "TickerProvider"
