+++ .editorconfig
@@ -1,0 +1,2 @@
+[*]
+end_of_line=lf
