--- Crycker.sln
@@ -1,30 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27004.2008
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Crycker", "Crycker\Crycker.csproj", "{36BF981F-E6CD-445D-B4F2-29012002EB7A}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{735ACC78-76AF-4CBF-8E34-1AA37E7981D5}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {82FBCA23-B22D-4E82-BA21-BAC3260F4597}
-	EndGlobalSection
-EndGlobal
