--- Crycker/Types/BoolEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Crycker.Types
-{
-    public class BoolEventArgs : EventArgs
-    {
-        public bool Value { get; private set; }
-
-        public BoolEventArgs(bool value)
-        {
-            Value = value;
-        }
-    }
-}
