--- Crycker/Types/IntEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Crycker.Types
-{
-    public class IntEventArgs : EventArgs
-    {
-        public int Value { get; private set; }
-
-        public IntEventArgs(int value)
-        {
-            Value = value;
-        }
-    }
-}
