--- Crycker/Types/StringEventArgs.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Crycker.Types
-{
-    public class StringEventArgs : EventArgs
-    {
-        public string Value { get; private set; }
-
-        public StringEventArgs(string value)
-        {
-            Value = value;
-        }
-    }
-}
