+++ CryptoTray.WinTray/cryptotray.ico
