+++ CryptoTray.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CryptoTray.WinTray", "CryptoTray.WinTray\CryptoTray.WinTray.csproj", "{36BF981F-E6CD-445D-B4F2-29012002EB7A}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{735ACC78-76AF-4CBF-8E34-1AA37E7981D5}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+		README.md = README.md
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CryptoTray", "CryptoTray\CryptoTray.csproj", "{D6EA3DEE-6D12-40F2-8162-9FC5A3B144EF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36BF981F-E6CD-445D-B4F2-29012002EB7A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D6EA3DEE-6D12-40F2-8162-9FC5A3B144EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D6EA3DEE-6D12-40F2-8162-9FC5A3B144EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D6EA3DEE-6D12-40F2-8162-9FC5A3B144EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D6EA3DEE-6D12-40F2-8162-9FC5A3B144EF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {82FBCA23-B22D-4E82-BA21-BAC3260F4597}
+	EndGlobalSection
+EndGlobal
