+++ CryptoTray/Data/ITickerProvider.cs
@@ -1,0 +1,9 @@
+﻿namespace CryptoTray.Data
+{
+    public interface ITickerProvider
+    {
+        string TickerName { get; }
+        string TickerUrl { get; }
+        string TickerKey { get; }
+    }
+}
