+++ CryptoTray/Types/BoolEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace CryptoTray.Types
+{
+    public class BoolEventArgs : EventArgs
+    {
+        public bool Value { get; private set; }
+
+        public BoolEventArgs(bool value)
+        {
+            Value = value;
+        }
+    }
+}
