+++ CryptoTray/Types/IntEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace CryptoTray.Types
+{
+    public class IntEventArgs : EventArgs
+    {
+        public int Value { get; private set; }
+
+        public IntEventArgs(int value)
+        {
+            Value = value;
+        }
+    }
+}
