+++ CryptoTray/Types/StringEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace CryptoTray.Types
+{
+    public class StringEventArgs : EventArgs
+    {
+        public string Value { get; private set; }
+
+        public StringEventArgs(string value)
+        {
+            Value = value;
+        }
+    }
+}
